\name{plm}
\alias{plm}
\title{A function to evaluates an associated Legendre polynomials}
\description{A function which numerically evaluates an associated Legendre polynomials for a 
given degree l and order m.}
\usage{
plm(l, m, x.entered)
}

\arguments{
  \item{l}{The degree of the associated Legendre polynomial. Must me a non negative integer.}
  \item{m}{The order of the associated Legendre polynomial. Must be an integer.}
  \item{x.entered}{Value to evaluate the Legendre polynomial.}
}
\details{}
\value{}
\references{}
\author{Nishan Mudalige}
\note{}
\seealso{}
\examples{
## plm(1,1,0.5)

## plm(4,-3,0.5)
}
