\name{s.harmonic-package}
\alias{s.harmonic-package}
\alias{s.harmonic}
\docType{package}
\title{
\packageTitle{s.harmonic}
}
\description{
\packageDescription{s.harmonic}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{s.harmonic}
\packageIndices{s.harmonic}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{s.harmonic}

Maintainer: \packageMaintainer{s.harmonic}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
